Abstract the PartAttachment class
Now "Attachments" are much easier to implement for different models
