+++ InvenTree/InvenTree/models.py
@@ -4,8 +4,11 @@ Generic models which provide extra functionality over base Django model types.
 
 from __future__ import unicode_literals
 
+import os
+
 from django.db import models
 from django.contrib.contenttypes.models import ContentType
+from django.utils.translation import gettext_lazy as _
 
 from django.db.models.signals import pre_delete
 from django.dispatch import receiver
@@ -13,6 +16,51 @@ from django.dispatch import receiver
 from mptt.models import MPTTModel, TreeForeignKey
 
 from .validators import validate_tree_name
+
+
+def rename_attachment(instance, filename):
+    """
+    Function for renaming an attachment file.
+    The subdirectory for the uploaded file is determined by the implementing class.
+
+        Args:
+        instance: Instance of a PartAttachment object
+        filename: name of uploaded file
+
+    Returns:
+        path to store file, format: '<subdir>/<id>/filename'
+    """
+
+    # Construct a path to store a file attachment for a given model type
+    return os.path.join(instance.getSubdir(), filename)
+
+
+class InvenTreeAttachment(models.Model):
+    """ Provides an abstracted class for managing file attachments.
+
+    Attributes:
+        attachment: File
+        comment: String descriptor for the attachment
+    """
+    def getSubdir(self):
+        """
+        Return the subdirectory under which attachments should be stored.
+        Note: Re-implement this for each subclass of InvenTreeAttachment
+        """
+
+        return "attachments"
+
+    attachment = models.FileField(upload_to=rename_attachment,
+                                  help_text=_('Select file to attach'))
+
+    comment = models.CharField(max_length=100, help_text=_('File comment'))
+
+    @property
+    def basename(self):
+        return os.path.basename(self.attachment.name)
+
+    class Meta:
+        abstract = True
 
 
 class InvenTreeTree(MPTTModel):
